Replace the old test infrastructure with the new Also make the appropriate changes to use it for ASAN tests as well. ASAN tests remain stubbed out while development on that sanitizer continues.
