+++ test/asan/asan_base.c
@@ -3,6 +3,9 @@
 
 int main() {
   int *nullptr = (void *)0x0;
+
+  // Stub this check while the ASAN runtime is being implemented
+  /* CHECK: .*? */
   *nullptr = 1;
 
   return 0;
