+++ test/test_runner.py
@@ -1,54 +1,59 @@
 #!/usr/bin/env python3
 import argparse
 import logging
-import json
-import os.path
 import subprocess
 import sys
 import re
 
 log = logging.getLogger("TestRunner")
 
+def extract_src_tests(filename):
+    test_markup = re.compile(r'/\*\s+CHECK:\s*(.*?)\s+\*/', re.MULTILINE | re.DOTALL)
+    with open(filename) as f:
+        src = f.read()
+        matches = test_markup.findall(src)
+        log.debug(f'test cases found: {matches}')
+        test_cases = [re.compile(x) for x in matches]
+        return test_cases
 
-def run_test(config, test):
-    print(f'-----------------------------------------------\n')
-    test_regexes = []
-    for test_case in config:
-        log.debug(f"Found test for {test_case}: {config[test_case]}")
-        test_regexes.append(re.compile(config[test_case]))
+def run_test(test):
+    print(f'-------------------------------------------------------')
 
-    test_regexes.reverse()
+    (src, exe) = test
+    test_cases = extract_src_tests(src)
 
-    log.info(f"running {test} {test_regexes}")
-    output = subprocess.run(test, capture_output=True)
-    log.debug(f"{test} output: [{output}]")
-
-    stderr_log = output.stderr.splitlines()
-    out_log = b'\n'.join(output.stdout.splitlines())
-    err_log = b'\n'.join(output.stderr.splitlines())
-    log.debug(f"\t\tstdout:\n\t{out_log}\n\t\tstderr:\n{err_log}")
-
-    if len(stderr_log) <= 0:
-        log.error(f'\t{test} failed to return an error')
-        print(f'-------------------------------------------------------\n\n')
+    if len(test_cases) <= 0:
+        log.error(f'No test cases found for {exe} in {src}')
         sys.exit(1)
 
-    for line in stderr_log:
-        if test_regexes is None or len(test_regexes) == 0:
-            print("skipping")
-            break
 
-        regex = test_regexes.pop()
-        result = regex.search(line.decode('ascii'))
+    log.info(f"running {exe} {test_cases}")
+    output = subprocess.run(exe, capture_output=True)
+    log.debug(f"{exe} output: [{output}]")
+
+    stderr_log = output.stderr.decode('ascii')
+    err_log = b"\n".join(output.stderr.splitlines())
+    log.info(f'test stderr:\n\n{err_log}\n')
+
+    if len(stderr_log) <= 0:
+        log.error(f'{test} failed to return an error')
+        print(f'-------------------------------------------------------')
+        sys.exit(1)
+
+    log.debug(f'test cases: {test_cases}')
+    for case in test_cases:
+        log.info(f'checking {case}')
+        result = re.search(case, stderr_log)
         log.debug(f'found {result}')
         if not result:
-            log.error(f"\t{test} failed")
-            log.error(f"\t\t{regex} does not match '{line}'")
-            print(f'-------------------------------------------------------\n\n')
+            log.error(f"test {test} failed")
+            print(f'-------------------------------------------------------')
             sys.exit(1)
-    log.info(f"\t{test} passed!")
-    print(f'-------------------------------------------------------\n\n')
+        else:
+            log.info('matched!')
 
+    log.info(f"\t{exe} passed!")
+    print(f'-------------------------------------------------------')
 
 def init_stdout_logger(loglevel=logging.INFO):
     root = logging.getLogger()
@@ -59,13 +64,11 @@ def init_stdout_logger(loglevel=logging.INFO):
     handler.setFormatter(formatter)
     root.addHandler(handler)
 
-
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument("-c", "--config", default="test_config.json", required=True)
-    parser.add_argument("-t", "--test", action="append", help="Run test executable")
+    parser.add_argument("-s", "--source", nargs='+', required=True, help="test source")
+    parser.add_argument("-t", "--test", nargs='+', required=True, help="test executable")
     parser.add_argument('--verbose', action='store_true')
-    parser.add_argument("test_list", nargs=argparse.REMAINDER)
 
     args = parser.parse_args()
 
@@ -74,36 +77,14 @@ def main():
     else:
         init_stdout_logger()
 
-    with open(args.config) as conf:
-        config = json.load(conf)
+    test_pairs = []
+    for src, exe in zip(args.source, args.test):
+        test_pairs.append((src, exe))
 
-    tests = []
-    if args.test is not None:
-        tests.extend(args.test)
-    if args.test_list is not None:
-        tests.extend(args.test_list)
-
-    for test in tests:
-        if not os.path.exists(test):
-            log.fatal(f"Could not find test '{test}'")
-            sys.exit(1)
-
-    log.debug(f"args {args}")
-    log.debug(f"tests {tests}")
-    log.debug(f"config {config}")
-
-    for test in tests:
-        base = os.path.basename(test)
-        subconfig = dict()
-        try:
-            subconfig = config[base]
-        except Exception as e:
-            log.fatal(f"Test config {args.config} missing entry for {test}: {e}")
-            sys.exit(1)
-        run_test(subconfig, test)
+    for test in test_pairs:
+        run_test(test)
 
     log.info("All tests passed")
-
 
 if __name__ == "__main__":
     main()
